requirements.txt
faster-whisper
pydub
flask
waitress
torch
torchaudio
dataoceanai-dolphin